general code improvements

+++ src/ChessCLIEngine/ChessCLIEngine/BitBoard.h
@@ -70,7 +70,7 @@ public:
 	BitBoard(std::string fen);
 	BitBoard(u64 pieces[SIDES][NUMBER_OF_PIECES], const AttackDictionary& attackPatterns, uint8_t flags, uint8_t enPassant);
 
-	float evaluate() const;
+	int evaluate() const;
 
 	std::shared_ptr<BitBoard> move(int startSquare, int endSquare, int promotionPiece=NO_PROMOTION) const;
 	std::shared_ptr<BitBoard> castleMove(bool isLong) const;
